What a fixture states is read by something that only reads fixtures
`ExampleVerifier` was 2709 lines, and its own class doc already named the rule that would have
kept it smaller. Two things were in it. One evaluates a source's rows and reports what did not
hold. The other reads a module's written statements against each other and reports where two of
them disagree. But the seam is not between those two: both sit on the same fixture builder —
`built`, `raw`, `decode`, `fixtureType`, `caseOnly`, `constructedCase`, `builtExpectedOrNull`,
`describeExpected` — instance methods over `module`, `symbols`, `values`, `helpers`, `neutral`,
`loader`, `bindings` and `expanding`, which is roughly the middle third of the file. Splitting the
two readers apart first would have copied that builder into one of them or left it reaching back
into the class it came from, and neither is a split.

So the thing that wanted a name is underneath both: something that answers what value this text
states, against this type, and answers nothing else. `FixtureReader` is that. It carries the module,
the symbols, the values a row may name, the loader, and the two collaborators a fixture needs — the
neutral form a decoder reads, and the invoker that runs a helper as the method its module emits —
along with the bindings in force and the values being expanded. It has no `Diagnostic`, no
`SourceRef`, no signature, no budget and no worker: what a fixture that will not build means, and
which report says so, differ between the row that runs it and the statement that is only read, and
each says it where it is.

Showing a value came with it, though showing is not reading. It needs the generated encoders, the
loader and the value that was built, and both readers above need it; left in either one, the other
would have grown a second copy of it.

This commit moves the methods and nothing else. No name changes, no message changes, no signature
changes, and the public surface is untouched — `ExampleVerifier` holds one reader and calls through
to it, so what each caller reads is what it read before. Which of those callers have to hold the
*same* reader, and which must not, is the next commit's question: today the isolation a timed-out
worker needs is still expressed by copying the whole verifier.

It is 1810 lines and 1000 lines, from 2709.

Refs #316
